Pass the blob URL for preloads in WasmEMCCBenchmark.
Right now we pass the original path when loading preloads for WasmEMCCBenchmark. This means we might skip the cache
and fetch the content from the network again. We don't want to do this because it means the OS might spin down the
CPU, which can punish running faster.

I also moved this logic to the `prerunCode` rather than adding it to the `runnerCode` since logically that
makes more sense.

As a drive by change, this patch has the preloads tuples contain the path for cli runs. This means we no longer have to duplicate
the preload paths into the benchmark.js file. Additionally, remove a console.log of the test that just finished which broke the
dumpJSONResults option for the CLIs.