StartupBenchmark should create sources lazily
Right now StartupBenchmark builds all the sources it will need up front.
This adds somewhat substantial memory overhead, in particular for
jsdom-d3-startup where it seems to add about 50MB of overhead.

This patch changes it so the source is generated in the
prepareForNextIteration hook on demand before the next iteration starts.